change commission to double, change age and exp to dateOfBirth and careerStartDate and change related classes

+++ src/main/java/com/task/test/footballmanager/dto/FootballClubDTO.java
@@ -16,7 +16,7 @@ public class FootballClubDTO {
     @NotNull
     private String name;
     @NotNull
-    private Integer commission;
+    private Double commission;
     @NotNull
     private BigDecimal balance;
 }
